# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/customers": {
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "summary": "Create a new customer",
        "tags": [
          "customers"
        ]
      },
      "get": {
        "operationId": "CustomersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of customers retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        },
        "summary": "Retrieve all customers",
        "tags": [
          "customers"
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "summary": "Retrieve a customer by ID",
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "summary": "Update an existing customer",
        "tags": [
          "customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "summary": "Delete a customer by ID",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/appointments": {
      "get": {
        "operationId": "AppointmentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all appointments retrieved successfully."
          }
        },
        "summary": "Get all appointments",
        "tags": [
          "appointments"
        ]
      },
      "post": {
        "operationId": "AppointmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        },
        "summary": "Create a new appointment",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/professional/{professionalId}": {
      "get": {
        "operationId": "AppointmentsController_findByProfessional",
        "parameters": [
          {
            "name": "professionalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments retrieved successfully."
          }
        },
        "summary": "Get all appointments for a professional",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/{id}": {
      "patch": {
        "operationId": "AppointmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated successfully."
          }
        },
        "summary": "Update an appointment",
        "tags": [
          "appointments"
        ]
      },
      "delete": {
        "operationId": "AppointmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment deleted successfully."
          }
        },
        "summary": "Delete an appointment",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/{id}/status": {
      "patch": {
        "operationId": "AppointmentsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment status updated successfully."
          }
        },
        "summary": "Update appointment status",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/check-availability": {
      "post": {
        "operationId": "AppointmentsController_checkAvailability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability check completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "available": {
                      "type": "boolean",
                      "description": "Whether the time slot is available"
                    },
                    "conflicts": {
                      "type": "object",
                      "properties": {
                        "appointments": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string"
                              },
                              "startDate": {
                                "type": "string"
                              },
                              "startTime": {
                                "type": "string"
                              },
                              "endDate": {
                                "type": "string"
                              },
                              "endTime": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid time range or professional is not active"
          },
          "404": {
            "description": "Professional not found"
          }
        },
        "summary": "Check appointment availability",
        "tags": [
          "appointments"
        ]
      }
    }
  },
  "info": {
    "title": "Backend Admin",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Full name of the Customer",
            "example": "John Doe"
          },
          "address": {
            "type": "string",
            "description": "Address of the Customer",
            "example": "123 Main St, City, Country"
          },
          "documents": {
            "type": "string",
            "description": "Customer documents (e.g., ID, insurance)",
            "example": "123456789"
          },
          "phone": {
            "type": "string",
            "description": "Customer phone number",
            "example": "+1234567890"
          },
          "email": {
            "type": "string",
            "description": "Customer email address",
            "example": "john.doe@example.com"
          },
          "communicationConsent": {
            "type": "boolean",
            "description": "Consent for communication according to LGPD",
            "example": true
          }
        },
        "required": [
          "fullName",
          "address",
          "documents",
          "phone",
          "communicationConsent"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {}
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {}
      },
...
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "customer": {
            "type": "string"
          },
          "professional": {
            "type": "string"
          },
          "schedule": {
            "type": "string"
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "example": "2025-05-02"
          },
          "startTime": {
            "type": "string",
            "example": "14:00"
          },
          "endDate": {
            "type": "string",
            "example": "2025-05-02"
          },
          "endTime": {
            "type": "string",
            "example": "15:00"
          }
        },
        "required": [
          "professional",
          "schedule",
          "title",
          "startDate",
          "startTime",
          "endDate",
          "endTime"
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {}
      },
      "UpdateAppointmentDto": {
        "type": "object",
        "properties": {
          "customer": {
            "type": "string"
          },
          "professional": {
            "type": "string"
          },
          "schedule": {
            "type": "string"
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        }
      },
      "UpdateAppointmentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status of the appointment",
            "enum": [
              "Agendado",
              "Pré-agendado",
              "Cancelado"
            ],
            "example": "Agendado"
          }
        },
        "required": [
          "status"
        ]
      },
      "CheckAvailabilityDto": {
        "type": "object",
        "properties": {
          "professional": {
            "description": "Professional ID",
            "example": "507f1f77bcf86cd799439011",
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "startDate": {
            "type": "string",
            "description": "Start date in YYYY-MM-DD format",
            "example": "2024-03-20"
          },
          "startTime": {
            "type": "string",
            "description": "Start time in HH:mm format",
            "example": "09:00"
          },
          "endDate": {
            "type": "string",
            "description": "End date in YYYY-MM-DD format",
            "example": "2024-03-20"
          },
          "endTime": {
            "type": "string",
            "description": "End time in HH:mm format",
            "example": "10:00"
          },
          "excludeId": {
            "description": "Appointment ID to exclude from check (for updates)",
            "example": "507f1f77bcf86cd799439011",
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          }
        },
        "required": [
          "professional",
          "startDate",
          "startTime",
          "endDate",
          "endTime"
        ]
      }
    }
  }
}